test_requirements.txt
coverage>=4.0.3,<5.0.0
django-nose>=1.4.3,<2.0.0
ddt>=1.0.1,<2.0.0
edx-django-oauth2-provider
edx-oauth2-provider
edx-lint==0.5.0
factory_boy>=2.6.1,<3.0.0
httpretty>=0.8.14,<1.0.0
mock>=1.3.0,<2.0.0
pep8>=1.7.0,<2.0.0
tox>=2.3.1,<3.0.0
git+https://github.com/edx/django-rest-framework.git@1ceda7c086fddffd1c440cc86856441bbf0bd9cb#egg=djangorestframework==3.6.3
pyjwkest==1.3.2
